% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chameleon.R
\name{pick_n_colors}
\alias{pick_n_colors}
\title{Pick a number of distinct colors which are sufficiently saturated.}
\usage{
pick_n_colors(n, minimal_saturation = 33)
}
\arguments{
\item{n}{The requested (positive) number of colors.}

\item{minimal_saturation}{Exclude colors whose saturation (`hypot(a, b)` in CIE-L*a*b* color space)
is less than this value (by default, 33).}
}
\value{
A list with two elements, `names` containing the color names and `lab` containing a
        matrix with a row per color and three columns containing the `l`, `a` and `b` coordinates
        of each color.
}
\description{
This ensures all colors are distinct by packing the (visible part) of the CIE-L*a*b* color space
with the needed number of spheres and using their centers to generate the colors.
}
\examples{
chameleon::pick_n_colors(8, 33)
}
